[train] refactor scaling policy resources
