+++ python/ray/train/v2/_internal/execution/scaling_policy/__init__.py
@@ -1,5 +1,10 @@
 # isort: off
 from .scaling_policy import ScalingDecision, ScalingPolicy, NoopDecision, ResizeDecision
+from .scaling_policy import (
+    AUTOSCALING_REQUESTS_EXPIRE_TIME_S,
+    AUTOSCALING_REQUESTS_GET_TIMEOUT_S,
+    AUTOSCALING_REQUESTS_INTERVAL_S,
+)
 from .elastic import ElasticScalingPolicy
 from .fixed import FixedScalingPolicy
 from .factory import create_scaling_policy
@@ -8,6 +13,9 @@ from .factory import create_scaling_policy
 
 
 __all__ = [
+    "AUTOSCALING_REQUESTS_EXPIRE_TIME_S",
+    "AUTOSCALING_REQUESTS_GET_TIMEOUT_S",
+    "AUTOSCALING_REQUESTS_INTERVAL_S",
     "ScalingPolicy",
     "ElasticScalingPolicy",
     "FixedScalingPolicy",
